introduce dedicated duration component
The inline duration component was rendering every second, changing the hooks and forcing a full re-render of the RHS every time. Move to a dedicated duration component that isolates the re-render to just that component, avoiding the user selector from re-rendering constantly and losing focus / preventing the use of the `No Assignee` feature.

Technically speaking, there is still the potential that a random re-render would temporarily break the component, and it's probably worth investigating that in the future.

Fixes: https://mattermost.atlassian.net/browse/MM-28032
Fixes: https://mattermost.atlassian.net/browse/MM-28036
